requirements.txt
streamlit
ultralytics
opencv-python-headless
streamlit-webrtc
numpy
Pillow
av